% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLPM.R
\name{.CLPM}
\alias{.CLPM}
\title{.CLPM}
\usage{
.CLPM(model, data, silent)
}
\arguments{
\item{model}{string specifying the model syntax}

\item{data}{longitudinal data in long format. Must have the columns "person"
and "occasion"}

\item{silent}{set to TRUE to suppress messages}
}
\value{
Model in RAM notation
}
\description{
creates a cross-laged panel model in RAM notation from a syntax similar to that
of lavaan.
}
\keyword{internal}
