% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLPM.R
\name{.fillCovariances}
\alias{.fillCovariances}
\title{.fillCovariances}
\usage{
.fillCovariances(S, latents, maxLag)
}
\arguments{
\item{S}{S matrix with undirected effects}

\item{latents}{data.frame with names of latent variables}

\item{maxLag}{larges lag in the equations}
}
\value{
updated S matrix
}
\description{
fills the covariances between latent variables which are considered initial covariances
}
\keyword{internal}
