% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLPM.R
\name{.fillIntercepts}
\alias{.fillIntercepts}
\title{.fillIntercepts}
\usage{
.fillIntercepts(M, latents, maxLag)
}
\arguments{
\item{M}{M matrix with means and intercepts}

\item{latents}{data.frame with names of latent variables}

\item{maxLag}{largest lag in the equations}
}
\value{
updated M matrix
}
\description{
fills the intercepts for latent variables which are considered initial intercepts
}
\keyword{internal}
