% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLPM.R
\name{.findMaxLag}
\alias{.findMaxLag}
\title{.findMaxLag}
\usage{
.findMaxLag(occasions)
}
\arguments{
\item{occasions}{data.frame with occasions}
}
\value{
integer; larges time lag
}
\description{
check the maximal time lag of the model
}
\keyword{internal}
