% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLPM.R
\name{.getVariableNamesCLPM}
\alias{.getVariableNamesCLPM}
\title{.getVariableNamesCLPM}
\usage{
.getVariableNamesCLPM(syntax)
}
\arguments{
\item{syntax}{string}
}
\value{
list with names of variables
}
\description{
extracts the names of the variables from the syntax
}
\keyword{internal}
