% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLPM.R
\name{.makeSingleLine}
\alias{.makeSingleLine}
\title{.makeSingleLine}
\usage{
.makeSingleLine(syntax)
}
\arguments{
\item{syntax}{string}
}
\value{
string with combined multi-line-statements
}
\description{
combine multi-line statements into one line
}
