% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAM2Lavaan.R
\name{.RAM2Lavaan}
\alias{.RAM2Lavaan}
\title{.RAM2Lavaan}
\usage{
.RAM2Lavaan(RAM, meanstructure)
}
\arguments{
\item{RAM}{model of type RAM}

\item{meanstructure}{boolean: should a mean structure be modeled?}
}
\value{
string to be used as model in lavaan
}
\description{
.RAM2Lavaan
}
\keyword{internal}
