% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLPM.R
\name{.removeWhitespace}
\alias{.removeWhitespace}
\title{.removeWhitespace}
\usage{
.removeWhitespace(syntax)
}
\arguments{
\item{syntax}{string}
}
\value{
string with removed whitespace
}
\description{
remove any whitespace from a string
}
