% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLPM.R
\name{.splitEquation}
\alias{.splitEquation}
\title{.splitEquation}
\usage{
.splitEquation(equation)
}
\arguments{
\item{equation}{string with equation}
}
\value{
data.frame with left hand side, right hand side, label, and operator
}
\description{
splits an equation in left hand side, right hand side, label, and operator
}
\keyword{internal}
