% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAM2Lavaan.R
\name{.valueOrLabel}
\alias{.valueOrLabel}
\title{.valueOrLabel}
\usage{
.valueOrLabel(RAM)
}
\arguments{
\item{RAM}{model of type RAM}
}
\value{
model of type RAM with distinction between value and label
}
\description{
separates between values and labels in RAM notation
}
\keyword{internal}
