% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{lessTemplates}
\alias{lessTemplates}
\alias{lessTemplates-package}
\title{lessTemplates}
\description{
The objective of lessTemplates is to provide a simplified implementation
for some common models which can be used in lessSEM. Currently included
are:
}
\details{
1.  cross-lagged panel models (with random intercepts).
2.  continuous time structural equation models.
3.  SEM with definition variables.

Additionally, lessSEM also provides some transformations for these
models which allow for regularizing specific model structures (e.g., to
test measurement invariance in cross-lagged panel models).

More details can be found here: https://github.com/jhorzek/lessTemplates
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jhorzek.github.io/lessTemplates/}
}

}
\author{
Jannik Orzek <orzek@mpib-berlin.mpg.de>
}
\keyword{internal}
