% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeARCL.R
\name{makeARCL}
\alias{makeARCL}
\title{makeARCL}
\usage{
makeARCL(
  data,
  latentIntercepts,
  ARCL,
  latentResiduals,
  loadings,
  manifestIntercepts,
  measurementError,
  lagSpecific = NULL
)
}
\arguments{
\item{data}{data set in long format. The following columns are required: (1) "id"
- a column which indicates which person a specific row of the data set belongs to.
(2) "occasion" - a column indicating which measurement occasion a row belongs to.
All other columns should contain the data required for the model. Optionally, a
column called "lag" can be added which specifies if parameters for this lag should
be estimated lag-specific (see below)}

\item{latentIntercepts}{nLatent x 1 matrix with latent intercepts. If a parameter
is fixed, set it to the numeric value (e.g., "0"). If a parameter is estimated, give it
a name (e.g., "latentIntercept1").}

\item{ARCL}{nLatent x nLatent matrix with autoregressive and cross-lagged parameters}

\item{latentResiduals}{nLatent x nLatent matrix with covariances of latent residuals}

\item{loadings}{an nManifest x nLatent matrix specifying the loadings.}

\item{manifestIntercepts}{nManifest x 1 matrix with manifest intercepts}

\item{measurementError}{nManifest x nManifest matrix with measurement errors}

\item{lagSpecific}{vector indicating which of the matrices above differs by lag.
Requires the specification of the lag column.}
}
\value{
list with lavaan model syntax and data set for lavaan
}
\description{
Creates and autoregressive cross-lagged panel model for use with lavaan or lessSEM.
}
\keyword{internal}
