% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RAM.R
\docType{class}
\name{RAM-class}
\alias{RAM-class}
\title{SEM in RAM notation}
\description{
SEM in RAM notation
}
\section{Slots}{

\describe{
\item{\code{A}}{character matrix with directed effects}

\item{\code{S}}{character matrix with undirected effects}

\item{\code{M}}{character matrix with intercepts}

\item{\code{F}}{numeric filter matrix}

\item{\code{latent}}{vector with names of latent variables}

\item{\code{manifest}}{vector with names of manifest variables}
}}

