% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{simplify}
\usage{
simplify(syntax)
}
\arguments{
\item{syntax}{model syntax from CLPM}
}
\value{
simplified model syntax
}
\description{
simplify a model syntax created by CLPM-function
}
