% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateCLPM}
\alias{simulateCLPM}
\title{simulateCLPM}
\usage{
simulateCLPM(seed = 123)
}
\arguments{
\item{seed}{seed used to simulate data}
}
\value{
data in long format
}
\description{
simulates data for a cross-lagged panel model with multiple indicators based on the syntax from
Jeroen D. Mulder & Ellen L. Hamaker (2021) Three Extensions of the Random
Intercept Cross-Lagged Panel Model, Structural Equation Modeling: A Multidisciplinary Journal,
28:4, 638-648, DOI: 10.1080/10705511.2020.1784738
}
\details{
see https://jeroendmulder.github.io/RI-CLPM/lavaan.html
}
