% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateExample1}
\alias{simulateExample1}
\title{simulateExample1}
\usage{
simulateExample1(seed = 123)
}
\arguments{
\item{seed}{seed used to simulate data}
}
\value{
list with data and model
}
\description{
Second order bivariate CLPM
}
