% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTSEMTransformations.R
\name{transformCTSEM}
\alias{transformCTSEM}
\title{transformCTSEM}
\usage{
transformCTSEM(CTSEM, parameters, transformation)
}
\arguments{
\item{CTSEM}{model created with CTSEM function}

\item{parameters}{names of the parameters which should be transformed. Note: Should be of form p(t)}

\item{transformation}{which transformation should be used? Currently supported: "changepoint", or "measurementInvariance"}
}
\value{
list with (1) transformation: string to be passed to lessSEM as transformation and (2) regularized: vector specifying which parameters should be regularized.
}
\description{
Transform the parameters of a CTSEM model. Returns a string with the
transformations to be passed to lessSEM.
}
\examples{
# TODO
}
